AF_XDP icmp echo.
For now it seems like xsk is a kernel bypass, meaning that packets
coming to our socket will not get back to kernel and we have to forward or drop them.

In contrast, source_xdp is a kernel hook, meaning that we get to look at
a packet and pass a verdict before the packet proceeds to kernel.

TODO: analyze a tradeof between these two. (check if AF_XDP has a better
performance and if the complexity of bypass is worth the performance)